Add api docs. Create BlockFactory.
BlockFactory ensures that block instances are re-used for interrelated files.
This makes sure that the equality checks work correctly for analysis and
optimization.

BlockFactory is not yet in use throughout the code or even yet added to
all the right public APIs where it will be needed. So this is just the
beginning of that refactor.