Refactor importing and Block Instantiation
The block importing system had some major issues:

  1. It was not cache-friendly which made it hard to avoid parsing the
     same block many times just to discover it would be the same block that
     was already parsed.
  2. It did not support source maps or dependent file tracking for precompilation.
  3. It did not provide a clean distinction between filenames and import
     identifiers, this created complexities for importers that wrap
     other importers.
  4. Because importing required a "from" file. Importing was useful for
     preprocessing files that were referenced but not for the blocks that are
     directly compiled. We needed a way to use the preprocessing consistently
     for primary load as well.

Additionally, block references would always result in a new Block
instance for a given compilation pass. Because block objects are
used as map keys for analysis and mapping, this would cause analyisis to
be incorrect when the same block was referenced multiple times.

To solve these issues, I have introduced a BlockFactory concept that
wraps importing to return a Block object for an entire compilation pass
that has a shared analysis. I have also changed importing from a
function with an extra property attached to it (which was hard to create
in typescript) to be an interface with methods that have much more well
defined semantics. There is a file identifier concept, which is just a
string, but it is intended to have an internal structure that the importer
understands and which may not be a file system path. This will allow
importers to clearly delegate to wrapped importers by annotating and
de-annotating the identifiers they use.

This is a backwards incompatible change.