Canonical identifiers.
In some cases an importer can consume and identifier that locates a file
but it is not the canonical identifier for that file. In that case, the
identifier returned for the imported file should be the canonical
identifier.

In other cases, a file can be imported from multiple importers with
different canonical identifier for each importer scheme. If those blocks
files have a filesystem representation we can ensure that the returned
block is always the same even if the importers are not coordinated --
this ensures consistency but it is less efficient and should be avoided
if at all possible.

Because canonical identifiers and real filenames for them cannot always
be known in advance, this patch updates the BlockFactory to track real
filenames as well as canonical identifiers and accounts for race
conditions in loading to always ensure that the blocks returned are the
correct first-returned unique instance. We do this without locking but
at the expense of sometimes loading or even parsing the same file
multiple times.