Path Alias Importer and Importer Tests.
We introduce a new importer, PathAliasImporter, that allows paths to be
aliased and reference with a developer-friendly prefix instead of always
being relative to the current file. The PathAliasImporter is a drop-in
replacement for the default filesystem identifier and will not break any
usage of the filesystem importer. In some cases, the PathAliasImporter
will return a different debug-fiendly value when an identifier is
inspected because the file falls within the scope of an aliased path
location.