Introduce a queue for preprocessing compilation.
The queue size can be set with ths "maxConcurrentCompiles" option which
defaults to 4.

When an error occurs, the queue is paused and all concurrent compilation
is completed before propagating the error.

This is especially important with precompilers that use native bindings
(I'm looking at you node-sass) because when an error propagates and
results in process.exit(), any pending native calls can cause the exit
to hang.