feat(analysis): style ordering independence.
This change lets styles be added to an element analysis in any order.
When the analysis is done, the analysis is sealed which causes the
analysis injest the styles in the correct order and prepare internal
caches for use.

Sealing is done automatically if the element analyis was being used
through the template analysis startElement/endElement methods but
must now be done explicitly otherwise.

Most methods on element analysis are only usable in either
unsealsed or sealed mode. This is probably indicative of
that this object is better represented as two different objects
(analyzer/analysis).