BREAKING CHANGE: Block no longer inherits from BlockObject.
Instead block.rootClass is used to get to a special BlockClass that is the root class of the block. All blockClass instances implement isRoot that will return whether it is the root style of the block.

BlockObject is no longer the type that is preferred for dealing with styles from a block generically, instead the type alias Style should be used.