feat: Simplify options.
BREAKING CHANGE: Some types related to options are removed/renamed.

The options code was some of the first I wrote for this project and it
was awkward to use. A lot of APIs expected a concrete options reader
instead of just the options interface in a read-only state.

The options reader is no longer exported, the new normalizeOptions
function should be used to get a read-only options object populated
with default values where necessary.

Where user-supplied options are accepted, the SparseOptions type is
accepted and then normalized.