feat: Display selector error locations using sourcemaps.
The source location displayed in error messages by default now will be
the source-mapped location for all selector errors.

In the CLI the generated source location as well as the source mapped
location are displayed. This is because, in my testing, the generated
source displays the error better and more reliably, but you still need a
pointer to the place in the authored source file to make the change,
especially in the case of block files that are composed of content from
multiple source files. Unfortunately, the sourcemap sometimes returns
a location that is not what the author would expect and not helpful.