feat: Per block namespaces.
RFC: https://github.com/linkedin/css-blocks/issues/332

BREAKING CHANGE

Ember template syntax for states now uses the block's name as the
namespace instead of the single namespace of 'state'. The default
namespace class attribute is now forbidden because the class attribute
is supposed to be namespaced like state attributes are. A new namespaced attribute
'scope' is used to apply the block root class to an element.