fix: Must invalidate the handlebars template cache across processes.
The CSS Blocks class name strategy uses a random hash that is
purposefully unstable across processes in order to prevent people
from relying on class names generated by css blocks. This means
that the persistent handlebars template cache must be invalidated
across builds.

Ideally this would only affect the templates that are ported to CSS
Blocks, but the ember plugin API for ast transforms lacks the expressive
power for that.

Because of this, we might want to reconsider the unstable classname
strategy in the future.