fix: Incorrect :scope selector and state output.
When a :scope selector was inserted, the class wasn't properly removed.
This was because the code that removed the class didn't always run.

When only a single state starting with "is-" belonged to a class, the
prefix wasn't removed. I fixed this by just removing prefixes much
earlier in the process rather than having it be part of the substate
algorithm.