fix: Create a 'Source Analysis' for the new rewrite strategy.
The ElementAnalysis gets sealed and a bunch of styles get added into the
mix based on composition and inheritance. This is ideal for conflict
detection and for sending the analysis to opticss, but it's not ideal
for our new rewrite strategy which will apply composition and
inheritance after resolving runtime blocks and only takes as input the
exact styles that have been applied by the source author.

In order to prevent existing code from breaking, I've created a new
analysis object called 'ElementSourceAnalysis' which is in charge of
creating an "unresolved" view of the analysis. This is used in the
new rewriter now.