refactor: Replace interface indices with style names.
Interface indices are unstable across changes to a css-block file which
makes the rewrite unstable across versions of dependency even when the
changes are only additive.

In order to simplify the rewrite and design, the style name is used
instead. In theory, we can optimize the names away in the
@css-blocks/ember-app build.