refactor: Remove block-interface-index.
Block Interface Index doesn't quite work as we originally hoped when
dealing with transitive dependencies. For now, we'll use node names for
interface indexing in the template runtime and work to reintroduce
numerical indexing later at app build time.

As a result of this, this commit rolls back the block-interface-index
decl in definition files.