fix: Only consider output styles that can possibly apply.
At service startup we construct a lookup map that maps from style ids to
possible optimized class entries.

Then for a given set of styles being applied to the current element, we
construct a list of the optimization entries that might evaluate to
true because the style expressions mention the input styles in some way.