feat: Helper invocation performance experiment.
Changes the way the css-blocks helper is invoked. Instead of passing
many arguments representing all the styles and their runtime state, a
single serialized JSON value is passed that represents all the styles.
the remainder of the arguments are the arguments passed by the author to
the various block attributes.