fix: Set eyeglass root within a directory-scoped processor.
In a directory scoped processor, we want to treat the root directory
of the processor as the eyeglass root. We override any root setting
that the application's configuration has provided, but we allow the
root to be overridden in the processor's setupOptions() method.

Usually the application's root is good enough for compilation, but there
are situation where the directory structures don't align and it's
important for the scoped processor to have it's root take precendence.
Namely: yarn link'ed development requires this.